Move SafeMultisigTransactionResponse type to safe-core-sdk-types
