+++ packages/safe-service-client/src/types/safeTransactionServiceTypes.ts
@@ -1,5 +1,8 @@
 import { Signer } from '@ethersproject/abstract-signer'
-import { SafeTransactionData } from '@safe-global/safe-core-sdk-types'
+import {
+  SafeMultisigTransactionResponse,
+  SafeTransactionData
+} from '@safe-global/safe-core-sdk-types'
 
 export type SafeServiceInfoResponse = {
   readonly name: string
@@ -97,22 +100,6 @@ export type SignatureResponse = {
   readonly signature: string
 }
 
-export type SafeMultisigConfirmationResponse = {
-  readonly owner: string
-  readonly submissionDate: string
-  readonly transactionHash?: string
-  readonly confirmationType?: string
-  readonly signature: string
-  readonly signatureType?: string
-}
-
-export type SafeMultisigConfirmationListResponse = {
-  readonly count: number
-  readonly next?: string
-  readonly previous?: string
-  readonly results: SafeMultisigConfirmationResponse[]
-}
-
 export type ProposeTransactionProps = {
   safeAddress: string
   safeTransactionData: SafeTransactionData
@@ -120,37 +107,6 @@ export type ProposeTransactionProps = {
   senderAddress: string
   senderSignature: string
   origin?: string
-}
-
-export type SafeMultisigTransactionResponse = {
-  readonly safe: string
-  readonly to: string
-  readonly value: string
-  readonly data?: string
-  readonly operation: number
-  readonly gasToken: string
-  readonly safeTxGas: number
-  readonly baseGas: number
-  readonly gasPrice: string
-  readonly refundReceiver?: string
-  readonly nonce: number
-  readonly executionDate: string
-  readonly submissionDate: string
-  readonly modified: string
-  readonly blockNumber?: number
-  readonly transactionHash: string
-  readonly safeTxHash: string
-  readonly executor?: string
-  readonly isExecuted: boolean
-  readonly isSuccessful?: boolean
-  readonly ethGasPrice?: string
-  readonly gasUsed?: number
-  readonly fee?: number
-  readonly origin: string
-  readonly dataDecoded?: string
-  readonly confirmationsRequired: number
-  readonly confirmations?: SafeMultisigConfirmationResponse[]
-  readonly signatures?: string
 }
 
 export type SafeMultisigTransactionListResponse = {
